third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/required_csp-header-expected.txt

This is a testharness.js-based test.
Found 70 tests; 48 PASS, 22 FAIL, 0 TIMEOUT, 0 NOTRUN.
PASS Test same origin: Sec-Required-CSP is not sent if `csp` attribute is not set on <iframe>.
PASS Test same origin redirect: Sec-Required-CSP is not sent if `csp` attribute is not set on <iframe>.
PASS Test cross origin redirect: Sec-Required-CSP is not sent if `csp` attribute is not set on <iframe>.
PASS Test cross origin redirect of cross origin iframe: Sec-Required-CSP is not sent if `csp` attribute is not set on <iframe>.
FAIL Test Required-CSP value on `csp` change: Sec-Required-CSP is not sent if `csp` attribute is not set on <iframe>. assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
FAIL Test same origin: Send Sec-Required-CSP when `csp` attribute of <iframe> is not empty. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test same origin redirect: Send Sec-Required-CSP when `csp` attribute of <iframe> is not empty. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test cross origin redirect: Send Sec-Required-CSP when `csp` attribute of <iframe> is not empty. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test cross origin redirect of cross origin iframe: Send Sec-Required-CSP when `csp` attribute of <iframe> is not empty. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test Required-CSP value on `csp` change: Send Sec-Required-CSP when `csp` attribute of <iframe> is not empty. assert_equals: expected (string) "script-src 'unsafe-inline'" but got (object) null
FAIL Test same origin: Send Sec-Required-CSP Header on change of `src` attribute on iframe. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test same origin redirect: Send Sec-Required-CSP Header on change of `src` attribute on iframe. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test cross origin redirect: Send Sec-Required-CSP Header on change of `src` attribute on iframe. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test cross origin redirect of cross origin iframe: Send Sec-Required-CSP Header on change of `src` attribute on iframe. assert_unreached: Child iframes have unexpected csp:"null" Reached unreachable code
FAIL Test Required-CSP value on `csp` change: Send Sec-Required-CSP Header on change of `src` attribute on iframe. assert_equals: expected (string) "script-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - gibberish csp
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - gibberish csp
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - gibberish csp
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - gibberish csp
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - gibberish csp assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name in multiple directives
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name in multiple directives
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name in multiple directives
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name in multiple directives
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - unknown policy name in multiple directives assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - misspeled 'none'
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - misspeled 'none'
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - misspeled 'none'
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - misspeled 'none'
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - misspeled 'none' assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - query values in path
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - query values in path
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - query values in path
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - query values in path
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - query values in path assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - missing semicolon
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - missing semicolon
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - missing semicolon
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - missing semicolon
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - missing semicolon assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - comma separated
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - comma separated
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - comma separated
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - comma separated
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - comma separated assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - html encoded string
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - html encoded string
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - html encoded string
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - html encoded string
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - html encoded string assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - url encoded string
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - url encoded string
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - url encoded string
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - url encoded string
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - url encoded string assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-uri present
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-uri present
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-uri present
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-uri present
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-uri present assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
PASS Test same origin: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-to present
PASS Test same origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-to present
PASS Test cross origin redirect: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-to present
PASS Test cross origin redirect of cross origin iframe: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-to present
FAIL Test Required-CSP value on `csp` change: Wrong value of `csp` should not trigger sending Sec-Required-CSP Header - report-to present assert_equals: expected (string) "default-src 'unsafe-inline'" but got (object) null
Harness: the test ran to completion.
